Massive refactor into multiple crates
