+++ pigdef/src/lib.rs
@@ -1,0 +1,7 @@
+//! This module provides definition structs for hardware, hardware events and configuration of hardware
+pub mod config;
+pub mod description;
+pub mod net_values;
+pub mod pin_function;
+#[cfg(feature = "usb")]
+pub mod usb_values;
