+++ pigdef/src/net_values.rs
