--- piggui/src/hw/mod.rs
@@ -1,5 +1,0 @@
-pub mod config;
-pub mod driver;
-mod hardware_description;
-mod pin_descriptions;
-mod pin_function;
